==> fecal_analysis/PhaseI_MB4,MB5,MB6/3. Baseline_Ext.Rmd <==
---
title: "R Notebook"
output: word_document
editor_options: 
  chunk_output_type: console
---

#To Update
```{r}
parent_location=c("t:DCEG/CGF/TechTransfer/Microbiome/Extraction/Optimization/Fecal/Fresh\ Fecal\ Optimization_2017.08/Phase\ I/Analysis/NP0084-MB4,5,6/R_Complete/")

output_location<-c("t:DCEG/CGF/TechTransfer/Microbiome/Extraction/Optimization/Fecal/Fresh\ Fecal\ Optimization_2017.08/Phase\ I/Analysis/NP0084-MB4,5,6/R_ExtVariation/")

taxa_level<-c("Genus")
reference_db<-"greengenes"
ext_choice<-"STD" #STD or HMO

manifest_name<-paste("metadata_",taxa_level,"_",reference_db,"_extcontrol_",ext_choice,".csv",sep="")
taxa_file<-paste("taxa_",taxa_level,"_",reference_db,"_extcontrol_",ext_choice,".csv",sep="")

control_location<-c("t:DCEG/CGF/TechTransfer/Microbiome/Extraction/Optimization/Fecal/Fresh Fecal Optimization_2017.08/Phase I/Controls/Taxonomy/")

worktable_name<-(paste(parent_location,"workspace_2020-01-05.RData",sep=""))
```

#Load required libraries, worktable
############### 
```{r}
library("ape")
library("biom")
library(cowplot)
library(data.table)
library(ggplot2)
library(grid)
library(gridExtra)
library("igraph")
library(qiime2R)
library(phyloseq)
library(scales)
library(structSSI)
library(tibble)
library(tidyverse)
library(tidyr)
library(tm)
library(vegan)
library(base)

load(worktable_name)
```

#Analysis
###############
#Directory Creation
```{r}
#Create directories
dir_create<-c("Data","OTUs","Taxa","Graphs")
for(a in dir_create){
  if (dir.exists(paste(output_location,a,sep=""))==FALSE){
    dir.create(paste(output_location,a,sep=""))
  }
}

#Move metadata, taxonomy,otu files
file.copy(paste(parent_location,"Taxa/",manifest_name,sep=""),paste(output_location,"/Data",sep=""))
file.copy(paste(parent_location,"Taxa/",taxa_file,sep=""),paste(output_location,"/Taxa",sep=""))
file.copy(paste(parent_location,"OTUs/otu_table.csv",sep=""),paste(output_location,"/OTUs",sep=""))
```

#Taxonomy Reference - Expected frequencies
```{r}
#Read in controls taxonomy file, grouped by taxa
for (a in taxa_level){
 control_expect <- read.table(paste(control_location,"Taxonomy_",taxa_level,"_2019.txt",sep=""),sep="\t",header=TRUE)
 rownames(control_expect) <- unique(control_expect[,taxa_level])
 
 control_db_name<-paste("expected_taxa_",a,sep="")
 assign(control_db_name,control_expect)
 remove(control_expect)
}

```

#Load OTU and metadata tables
```{r}
for(a in taxa_level){
 for (b in reference_db){
  metadata_db_name<-paste("metadata_",a,"_",b,sep="")
  
  temp <- read.table(paste(output_location,"Data/",metadata_db_name,"_extcontrol_",ext_choice,".csv",sep=""),sep=",",header= TRUE)
  rownames(temp)<-temp$X
  temp<-temp[-1]
  
  assign(metadata_db_name,temp)
  remove(temp)
  
  taxa_db_name<-paste("taxa_",a,"_",b,sep="")
  
  temp<-read.table(paste(output_location,"Taxa/",taxa_db_name,"_extcontrol_",ext_choice,".csv",sep=""),sep=",",header= TRUE)
  rownames(temp)<-temp$X
  temp<-temp[-1]
  
  assign(taxa_db_name,temp)
  remove(temp)
 }
}
```

#Load phylo object
```{r}
#NOTE: must run 1. Initial processing
#physeq_filt includes all samples, after filtering

#Subset to create standard ext only
subset_list<-c("Ext.Control") 
physeq_scale_sub<-subset_samples(physeq_filt, Sample.Type %in% subset_list)
 
if(ext_choice=="STD"){

 subset_list<-unique((sample_data(physeq_scale_sub)$Reciept))
 subset_list<-subset_list[1:22]
 physeq_scale_sub<-subset_samples(physeq_scale_sub, Reciept %in% subset_list)
} else{
 #Subset to create homo ext only
 subset_list<-unique((sample_data(physeq_scale_sub)$Reciept))
 subset_list<-subset_list[23:34]
 physeq_scale_sub<-subset_samples(physeq_scale_sub, Reciept %in% subset_list)
}
```

#OTU Counts by sample
#Determine total OTUs
```{r}

taxa_Genus_greengenes$Total_OTUs <- rowSums(taxa_Genus_greengenes)
```

#TAR/TDR Sample Level
```{r}
#Testing
# output<-tar_tdr
# db<-"greengenes"
# tl<-"Genus"
# ct<-"DZ35316"
# otherfilt<-"MagMax Microbiome Ultra Kit"
# remove(output,db,tl,ct,otherfilt)
# remove(sample_list,exp_taxa_list,unexp_taxa_list)

tp_fp_fn_samplelevel<- function(output,db,tl,ct,otherfilt){
 sample_list<-intersect(rownames(get(paste("taxa",tl,db,sep="_"))),rownames(subset(get(paste("metadata",tl,db,sep="_")),Source.Descrip==ct)))
 sample_list<-intersect(sample_list,rownames(subset(get(paste("metadata",tl,db,sep="_")),Ext.Kit==otherfilt)))#add another layer of filtering based on extraction kit
 
 exp_taxa_list<-rownames(get(paste("expected_taxa",tl,sep="_"))[ get(paste("expected_taxa",tl,sep="_"))[[ct]] !='A' , ])
 unexp_taxa_list<-setdiff(colnames(get(paste("taxa",tl,db,sep="_"))),exp_taxa_list)
 unexp_taxa_list<-unexp_taxa_list[-length(unexp_taxa_list)] #Remove Total_OTUs column
 unexp_taxa_list<-unexp_taxa_list[-45]#Remove Higher Genus
 
 for (a in sample_list){
  fp=0
  fn=0
  tp=0
  
  taxa_db<-get(paste("taxa",tl,db,sep="_"))
  
  #Run through expected taxa list
  for (b in exp_taxa_list){
   otu_count<-taxa_db[a,b]

   
   if(is.null(otu_count) || otu_count==0 || is.na(otu_count)){
    fn=fn+1

   } else if (otu_count>0){
    tp=tp+1
   }
  }
  
  #Run through all taxa not expected
  for (b in unexp_taxa_list){
   otu_count<-taxa_db[a,b]
   
   if(otu_count==0){
    next;
   } else if (otu_count>0){
    fp=fp+1
   }
  }
  
  output[a,"ControlType"]<-ct
  output[a,"Kit"]<-get(paste("metadata",tl,db,sep="_"))[a,"Ext.Kit"]
  output[a,"Expected"]<-length(exp_taxa_list)
  output[a,paste("TP",db,tl,sep="_")]<-tp
  output[a,paste("FP",db,tl,sep="_")]<-fp
  output[a,paste("FN",db,tl,sep="_")]<-fn
  output[a,paste("TAR",db,tl,sep="_")]<-tp/(tp+fp)
  output[a,paste("TDR",db,tl,sep="_")]<-tp/(tp+fn)
  output[a,paste("F1",db,tl,sep="_")]<-2*((tp/(tp+fp))*(tp/(tp+fn)))/((tp/(tp+fp))+(tp/(tp+fn)))
  output[a,"Homo.Method"]<-get(paste("metadata",tl,db,sep="_"))[a,"Homo.Method"]
 }
 
 return(output)
}

#Determine which controls
control_type<-unique(sample_data(physeq_scale_sub)$Source.Descrip)
control_type<-control_type[-c(3,4)] #Remove Water and Robogut
control_type

other_filt<-unique(sample_data(physeq_scale_sub)$Ext.Kit)


tar_tdr<-data.frame()
for(a in reference_db){
 for (b in taxa_level){
  for (c in control_type){
   for (d in other_filt){
    tar_tdr<-tp_fp_fn_samplelevel(tar_tdr,a,b,c,d) #D: change subsetting of sample_list as needed within funciton IE Source.Descrip or Ext.Kit
   }
  }
 }
}

write.csv(tar_tdr,paste(output_location,"\\Taxa\\Taxonomy_tar_trd_ext_",ext_choice,".csv",sep=""))

```

#TAR/TDR Averaged
```{r}
#Testing
# db<-subset(tar_tdr,Kit=="DSP Virus")
# opts<-"DSP Virus"
# db<-tar_tdr
# opts<-"HOMO"
# remove(db,opts,temp,count)

#update cols for kit inclusion
tar_tdr_av_kit<-function(db,opts,opts2){
 
 temp<-aggregate(db[, 3], list(db$ControlType), mean) #Expected tax
 row.names(temp)<-temp$Group.1
 temp<-temp[,-1]
 
 temp<-cbind(temp,aggregate(db[, 7], list(db$ControlType), mean)) #TAR Average
 row.names(temp)<-temp$Group.1
 temp<-temp[,-2]
 
 temp<-cbind(temp,aggregate(db[, 8], list(db$ControlType), mean)) #TDR Average
 row.names(temp)<-temp$Group.1
 temp<-temp[,-3]
 
 temp<-cbind(temp,aggregate(db[, 9], list(db$ControlType), mean)) #F1 Average
 temp<-temp[,-4]
 
 colnames(temp)<-c("Expected","TAR","TDR","F1Score")
 
 if(opts=="STD"){
  
 } else{
  count=1
  
  for(a in rownames(temp)){
   temp[a,"Kit"]<-opts
   temp[a,"Control"]<-a
   temp[a,"Homo.Method"]<-opts2
   rownames(temp)[count]<-paste(a,opts,opts2,sep="_")
   count=count+1
  }
 }

 return(temp)
}

tar_tdr_av<-data.frame()
if(ext_choice=="STD"){
 #Run for averages
 tar_tdr_av<-tar_tdr_av_kit(tar_tdr,ext_choice) 
 write.csv(tar_tdr_av,paste(output_location,"\\Taxa\\Taxonomy_tar_trd_ext_av.csv",sep=""))
 
 #Run by Kit
 tar_tdr_av<-data.frame("Expected"=1,"TAR"=1,"TDR"=1,"F1Score"=1,"Kit"=1,"Control"=1,"Homo.Method"=1)
 filt_list<-unique(tar_tdr$Kit)
 
 for(a in filt_list){
  temp<-subset(tar_tdr,Kit==a)
 
  tar_tdr_av<-rbind(tar_tdr_av,tar_tdr_av_kit(temp,a,""))
  remove(temp)
 }
 
 tar_tdr_av<-tar_tdr_av[-1,]#Remove starter row
 write.csv(tar_tdr_av,paste(output_location,"\\Taxa\\Taxonomy_tar_trd_ext_av_std_kit.csv",sep=""))
 
 } else{

 #Run by Homog
 tar_tdr_av<-data.frame("Expected"=1,"TAR"=1,"TDR"=1,"F1Score"=1,"Kit"=1,"Control"=1, "Homo.Method"=1)
 filt_list<-unique(tar_tdr$Kit)
 filt_list1<-unique(tar_tdr$Homo.Method)

 for (a in filt_list1){
  temp<-subset(tar_tdr,Homo.Method==a)
 
  tar_tdr_av<-rbind(tar_tdr_av,tar_tdr_av_kit(temp,"",a)) #db, kit, homo
  #remove(temp)
 }
 
 tar_tdr_av<-tar_tdr_av[-1,]#Remove starter row
 write.csv(tar_tdr_av,paste(output_location,"\\Taxa\\Taxonomy_tar_trd_ext_av_homo.csv",sep=""))
 
 #Run by homog, by kit
 tar_tdr_av<-data.frame("Expected"=1,"TAR"=1,"TDR"=1,"F1Score"=1,"Kit"=1,"Control"=1, "Homo.Method"=1)
 for(a in filt_list){
  temp<-subset(tar_tdr,Kit==a)
  for (b in filt_list1){
   temp1<-subset(temp,Homo.Method==b)
   print(a)
   print(b)
   if(nrow(temp)>0){
    print ("OK")
    tar_tdr_av<-rbind(tar_tdr_av,tar_tdr_av_kit(temp,a,b))
   }
  }
 }
 
 tar_tdr_av<-tar_tdr_av[-1,]#Remove starter row
 write.csv(tar_tdr_av,paste(output_location,"\\Taxa\\Taxonomy_tar_trd_ext_av_homo_kit.csv",sep=""))
 #remove(temp)
}


```

#Relative Abundance  - Ext Controls
```{r}
#Function for relative abundance - sample level
relabun<-function(taxadb,exptaxadb,metadb){
  #Testing
  # taxadb<-taxa_Genus_greengenes
  # exptaxadb<-expected_taxa_Genus
  # metadb<-metadata_Genus_greengenes
  # remove(taxadb,exptaxadb,metadb)
  
  relabun_extcont<-data.frame()

  #Complete Relative Abundance
  sample_list_extcont<-rownames(taxadb)
  taxa_list <- colnames(taxadb)
  taxa_list<-taxa_list[-(length(taxa_list))] #Remove total OTU's from list
  taxa_list<-unique(append(taxa_list,rownames(exptaxadb)))
  
  for (a in sample_list_extcont){
   relabun_extcont[a,"Source.Descrip"]<-as.character(metadb[a,"Source.Descrip"])
   relabun_extcont[a,"Ext.Kit"]<-as.character(metadb[a,"Ext.Kit"])
   relabun_extcont[a,"Homo.Method"]<-as.character(metadb[a,"Homo.Method"])
   
   for (b in taxa_list){
    relabun <- taxadb[a,b]
    
    if(!(is.na(relabun) || is.null(relabun) || relabun==0)){
      relabun_extcont[a,b] <- taxadb[a,b]/taxadb[a,"Total_OTUs"]
    } else{
     relabun_extcont[a,b] <- 0
    }
   }
  }
  
  write.csv(relabun_extcont,paste(output_location,"Data/relabun_extcont_",ext_choice,".csv",sep=""))
  return(relabun_extcont)
}

#Function for relative abundance - averaged
relabun_av<-function(db,taxadb,exptaxadb,type,subset){
  # #Testing
  # db<-relabun_extcont
  # subset<-"Ext.Kit"
  # remove(db,subset,temp,temp2,tempcols,temp3)

  #Determine unique subset list
  output_db<-data.frame()
  
  #Subset database by control
  for (a in unique(db[,"Source.Descrip"])){
    temp<-subset(db,Source.Descrip==a)
    
    for(b in unique(temp[,subset])){
      temp2<-temp[temp[,subset]==b,]
      tempcols<-colnames(temp2[,4:ncol(temp2)])
   
      #if the database includes this control
      taxa_list <- colnames(taxadb)
      taxa_list<-taxa_list[-(length(taxa_list))] #Remove total OTU's from list
      taxa_list<-unique(append(taxa_list,rownames(exptaxadb)))
      
      if(nrow(temp2)>0){
        rowcount=nrow(output_db)+1
        for (c in taxa_list){
          temp3<-mean(temp2[,c])
          output_db[rowcount,"Name"]<-paste(a,b,sep="_")
          output_db[rowcount,"Subset"]<-b
          output_db[rowcount,"Source.Descrip"]<-a
          output_db[rowcount,c]<-temp3
        } 
      }
    }
    
    rowcount=nrow(output_db)+1
    if(a!="Water"){
      for (b in rownames(exptaxadb)){
        output_db[rowcount,"Name"]<-paste(a,"expected",sep="_")
        output_db[rowcount,b]<-as.numeric(as.character(exptaxadb[b,a]))
      } 
    }
  }
  
  output_db[output_db=="A" ]=0 #replace A's with 0
  output_db[is.na(output_db)]=0 #replace A's with 0
  
  rownames(output_db)<-output_db$Name
  output_db<-output_db[,-1]
  write.csv(output_db,paste(output_location,"Data/relabun_extcont_av_",type,"_",subset,".csv",sep=""))
  return(output_db)
}
 
#Run functions
relabun_extcont<-relabun(taxa_Genus_greengenes,expected_taxa_Genus,metadata_Genus_greengenes)
relabun_av(relabun_extcont,taxa_Genus_greengenes,expected_taxa_Genus,"std","Source.Descrip")
relabun_extcont_av<-relabun_av(relabun_extcont,taxa_Genus_greengenes,expected_taxa_Genus,"std","Ext.Kit")
```

#Plot Rel Abun
```{r}
relabun_av_plot<-function(db,type){
  #Testing
  #db<-relabun_extcont_av
  #remove(relabun_extcont_av_plot,db)
  
  output_db<-data.frame()
  
  for (a in unique(db$Source.Descrip)[unique(db$Source.Descrip)!="Water"&unique(db$Source.Descrip)!="0"]){
    for (b in rownames(expected_taxa_Genus[expected_taxa_Genus[,a]!='A',])){
      for (c in unique(relabun_extcont$Ext.Kit)){
        rowcount<-nrow(output_db)+1
        output_db[rowcount,"Exp"]<-as.numeric(as.character(expected_taxa_Genus[b,a]))
        output_db[rowcount,"Obs"]<-db[paste(a,c,sep="_"),b]
        output_db[rowcount,"%Dif"]<-(output_db[rowcount,"Exp"]-output_db[rowcount,"Obs"])/output_db[rowcount,"Exp"]
        output_db[rowcount,"Tax"]<-b
        output_db[rowcount,"Source.Descrip"]<-a
        output_db[rowcount,"Ext.Kit"]<-c
      }
    }
  }
  output_db[is.na(output_db)]=0 #replace A's with 0
  write.csv(output_db,paste(output_location,"Data/relabun_extcont_av_",type,"_plot.csv",sep=""))
  return(output_db)
}

relabun_extcont_av_plot<-relabun_av_plot(relabun_extcont_av,"std")

```

#Cosine Similarity
```{r}
cosine_sim<-function(db,subset,type){
  
  output_db<-data.frame()
  
  for (a in unique(db$Source.Descrip)){
    temp<-db[db[,"Source.Descrip"]==a,]
    
    for (b in unique(temp[,subset])){
      temp2<-temp[temp[,subset]==b,]
      
      output_db[nrow(output_db)+1,"CosSim"]<-as.numeric(as.character(cosine(temp2$Obs,temp2$Exp)))
      output_db[nrow(output_db),subset]<-b
      output_db[nrow(output_db),"Source.Descrip"]<-a
    }
  }
  write.csv(output_db,paste(output_location,"Data/cosine_sim_",type,"_",subset,".csv",sep=""))
  output_db
}

cosine_sim(relabun_extcont_av_plot,"Ext.Kit","std")
```

#8 Alpha Diversity of Homo Methods
```{r}
otu<-otu_table(physeq_filt_extcontrols)
tax<-tax_table(physeq_filt_extcontrols)
random_tree<-random_tree<-rtree(ntaxa(physeq_filt_extcontrols), rooted=TRUE, tip.label=taxa_names(physeq_filt_extcontrols))
sdata<- sample_data(metadata_baseline)
temp<-phyloseq(otu,tax,sdata,random_tree)

# Scale reads to even depth 
physeq_scale_temp<-temp %>%
  scale_reads(n=1000) 

remove(temp)

#Homo Method
p<-plot_richness(physeq_scale_temp, x="Homo.Method") + geom_boxplot()
file_name =paste(project_location,"\\Output\\R\\Graphs\\Summary\\alphadiv_homomethod.tiff",sep="")
tiff(file_name, width=800)
print(p)
dev.off()

remove(physeq_scale_temp)
```
################
#Unused, but useful?

#Reproducibility - Original vs Replicate
```{r}
#https://cran.r-project.org/web/packages/qualityTools/vignettes/qualityTools.pdf

#Create genus list of Zymo Ext Control only
genus_temp <- control_expect[!is.na(control_expect[,"Zymo.Ext"]),]
genus_temp <- (genus_temp[,"Genus"])

#Create dataframe
temp<-data.frame()
count=1

#Match originals back to residuals 
#SC249391.PC04924.D03 & SC249391.PC04925.H01 to SC261522.PC04925.F11 
#SC304927.PC07578.B02 to SC304937.PC07578.D04
# SC304924.PC07578.B.07.PC07578.B07 to SC304934.PC07578.D.02.PC07578.D02 & SC304934.PC07578.C.05.PC07578.C05
sample_list <- c("SC249391.PC04925.H01", "SC261522.PC04925.F11", "SC304927.PC07578.B02", "SC304937.PC07578.D04", "SC304924.PC07578.B.07.PC07578.B07", "SC304934.PC07578.D.02.PC07578.D02" )

#For all genus
for (a in genus_temp){
  
 #Move through list
 for (b in sample_list){
   temp[count,"Genus"]<-a
   temp[count,"Obs_Per"]<- samples_obs_abund[a,b]
   temp[count,"Sample"]<-b
   count=count+1
 }
}

#Remove rows with NA
temp<- temp %>% drop_na()

###Testing
write.csv(temp,"repo.csv")    
temp <- read.csv("repo.csv",header = TRUE)

#Design Repro features
repro_resvsori <- gageRRDesign(Operators = 2, 
             Parts = length(unique(temp$Genus)), 
             Measurements = length(unique(temp$Sample))/2,
             randomize = FALSE)

#Set the response - must be numeric value
response(repro_resvsori)<- as.numeric(as.list(temp[,"Obs_Per"]))

#Run summary
repro_complete <- gageRR(repro_resvsori, method = "crossed", sigma =6, alpha=.25)
summary(repro_complete)

#Plot results
#Operator A = Original, B = Residual
#Parts = Genus (7 total, 8th was removed due to NA)
#Measurements - Samples (4 total per operator)
plot(repro_complete) 
```

#Reproducibilty - Extraction Kits
```{r}
#Create dataframe
temp<-data.frame()
count=1
for (a in control_types){
 #Create genus list of specfic control
 genus_temp <- control_expect[!is.na(control_expect[,a]),]
 genus_temp <- (genus_temp[,"Genus"])
 
 for (b in kit_types){
  
  ###Create database
  temp2 <- subset(samples_variables_f,Kit==b & Type==a & O.R == "Original")
  sample_list <- unique(temp2$NephID)

  #For each of the samples in the sample list
  for(c in sample_list){
   
   #For each Genus in list, determine the samples value
   for (d in genus_temp){
    temp[count,"Genus"]<-d
    temp[count,"Obs_Per"]<- samples_obs_abund[d,c]
    temp[count,"Sample"]<-c
    temp[count,"Kit"]<-b
    temp[count,"Type"]<-a
    count=count+1
   }
  }
 }
}

###Testing
write.csv(temp,"repo2.csv")    

#Load in final file (manipulated to choose six samples)
temp<- read.csv("Data/repo_ze_final1.csv",header = TRUE)

#Design Repro features
repro_resvsori <- gageRRDesign(Operators = 5, 
             Parts = length(unique(temp$Genus)), 
             Measurements = length(unique(temp$Sample))/5,
             randomize = FALSE)

#Set the response - must be numeric value
response(repro_resvsori)<- as.numeric(as.list(temp[,"Obs_Per"]))

#Run summary
repro_complete <- gageRR(repro_resvsori, method = "crossed", sigma =6, alpha=.25)
summary(repro_complete)

#Plot results
#Operator A = Qiagen DSP Virus, B = Zymo MagBead DNone Kit, C=MagAttract PowerMag Soil, D = MagAttract PowerMag Microbiome, E = Qiagen QIAamp  Modified
#Parts = Genus (8 total)
#Measurements - Samples (6 total per operator)
plot(repro_complete) 

```
